==> 01. Preprocessing/00. FirstName.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import warnings
warnings.filterwarnings('ignore')
%matplotlib inline

df = pd.read_csv('./car/tucson_engColumns.csv')
df

df['name'][1][3:]

df['name'] = df['name'].apply(lambda x : x[3:])
df

' '.join(df['name'][1900].split(' ')[:3])

df['name'].unique()

def first_name(name) :
    for i in range(len(name.split(' '))) :
        if 'ix' in name.split(' ')[i] :
            tmp = ' '.join(name.split(' ')[:i+1])
        elif '투싼' in name.split(' ')[i] :
            tmp = ' '.join(name.split(' ')[:i+1])
    return tmp

df['first_name'] = df['name'].apply(lambda x : first_name(x))
df

df.columns

df = df[['name', 'first_name', 'year', 'fuel', 'distance', 'color', 'sunroof', 'HID',
       'Al_wheel', 'auto_mirror', 'DLR', 'sunroof_pano', 'roof_rack', 'LED',
       'HBA', 'ghost_door', 'sliding_door', 'side_step', 'adap_h',
       'driver_heated_seat', 'driver_ventilated_seat', 'power_steering',
       'steering_remote', 'leahter_seat', 'driver_elec_seat', 'pass_elec_seat',
       'back_elec_seat', 'driver_memory_seat', 'massage_seat',
       'supervision_cluster', 'paddle_shift', 'hi_pass', 'pass_memory_seat',
       'back_heated_seat', 'amb_light', 'walkin_seat', 'sun_visor',
       'pass_ventilated_seat', 'back_ventilated_seat', 'back_sensor',
       'side_airbag', 'driver_airbag', 'pass_airbag', 'back_camera',
       'around_view', 'blackbox', 'ABS', 'ECS', 'TCS', 'ESC', 'LDWS',
       'bur_alarm', 'TPMS', 'knee_airbag', 'parking_sensor', 'front_carmera',
       'AEB', 'smartkey', 'power_window', 'auto_doorlock', 'fullauto_aircon',
       'auto_light', 'auto_parking', 'elec_parking_break', 'HUD', 'ECM',
       'cruise_control', 'auto_trunk', 'heated_steering', 'keyless_doorlock',
       'rain_sensor', 'stopNgo', 'wireless_charge', 'navi', 'CDrom',
       'CD_changer', 'av_sys', 'mp3', 'AUX', 'USB', 'handsfree', 'iPod',
       'back_monitor', 'mirroring', 'non_genu_navi', 'bluetooth', 'price'
       ]]

df

df['first_name'].value_counts()

set(df['first_name'])

set(df[df['first_name'] == '뉴모닝 SLX']['fuel'])

mapper = {'뉴 투싼 ix' : 'new_ix', 
          '올 뉴 투싼' : 'all_new', 
          '투싼ix' : 'ix'}

df['first_name'] = df['first_name'].map(mapper)
df

df = df[df['price'] < 8000].reset_index(drop=True)

df

df['first_name'].value_counts()

df.to_csv('./car/tucson_firstName.csv', index=False)
