==> 01. Preprocessing/01. Preprocessing_DataFrame.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings('ignore')
%matplotlib inline

import matplotlib.font_manager as fm
font_path = 'C:\\Windows\\Fonts\\gulim.ttc'
font_name = fm.FontProperties(fname=font_path).get_name()
plt.rc('font', family=font_name)

df_01 = pd.read_csv('./csv/grandeur/g01.csv')
df_02 = pd.read_csv('./csv/grandeur/g02.csv')
df_03 = pd.read_csv('./csv/grandeur/g03.csv')
df_04 = pd.read_csv('./csv/grandeur/g04.csv')
df_05 = pd.read_csv('./csv/grandeur/g05.csv')
# df_06 = pd.read_csv('./csv/grandeur/g06.csv')
# df_07 = pd.read_csv('./csv/grandeur/g07.csv')

df = pd.concat([df_01, df_02, df_03, df_04, df_05]).reset_index(drop=True)
df

df = df[df.duplicated() == False].reset_index(drop=True)
df

df['fuel'].unique()

df['fuel'].value_counts()

df[df['fuel'] == '가솔린+LPG']

df['name'].value_counts()

df.info()

df.isna().sum().sum()

df['name'].nunique()

df['color'].value_counts()

df['color'].unique()

df[df['color'].str.contains('베이지') | df['color'].str.contains('진주') | df['color'].str.contains('미색')
  | df['color'].str.contains('밀크')]

df[df['color'].str.contains('흰')]

df = df[df['price'] < 8000].reset_index(drop=True)
df.to_csv('./car/grandeur.csv', index=False)

df = pd.read_csv('./car/tucson.csv')
df

df = df[df['price'] < 8000].reset_index(drop=True)
df.to_csv('./car/tucson.csv', index=False)
