==> 01. Preprocessing/02. Preprocessing_Columns.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
import os
warnings.filterwarnings('ignore')
%matplotlib inline

os.listdir('./car')[0][:-4]

mapper = {
    '썬루프' : 'sunroof', 
    '제논라이트(HID)' : 'HID',
    '알루미늄휠' : 'Al_wheel', 
    '전동접이식미러' : 'auto_mirror', 
    '주간주행등(DLR)' : 'DLR', 
    '파노라마썬루프' : 'sunroof_pano', 
    '루프랙' : 'roof_rack', 
    'LED리어램프' : 'LED',
    '하이빔 어시스트' : 'HBA', 
    '고스트 도어 클로징' : 'ghost_door', 
    '자동슬라이딩도어' : 'sliding_door', 
    '전동사이드스탭' : 'side_step', 
    '어댑티드헤드램프' : 'adap_h', 
    '열선시트' : 'driver_heated_seat',
    '통풍시트' : 'driver_ventilated_seat', 
    '파워핸들' : 'power_steering', 
    '핸들 리모컨' : 'steering_remote', 
    '가죽시트' : 'leahter_seat',
    '운전석전동시트' : 'driver_elec_seat', 
    '조수석전동시트' : 'pass_elec_seat', 
    '뒷좌석전동시트' : 'back_elec_seat',
    '메모리시트' : 'driver_memory_seat',
    '안마시트' : 'massage_seat', 
    '슈퍼비전계기판' : 'supervision_cluster', 
    '패들시프트' : 'paddle_shift', 
    '하이패스' : 'hi_pass', 
    '메모리시트(동승석)' : 'pass_memory_seat', 
    '열선시트(뒷좌석)' : 'back_heated_seat',
    '엠비언트라이트' : 'amb_light',
    '워크인시트' : 'walkin_seat', 
    '전동햇빛가리개' : 'sun_visor', 
    '통풍시트(동승석)' : 'pass_ventilated_seat',
    '통풍시트(뒷좌석)' : 'back_ventilated_seat', 
    '후방감지센서' : 'back_sensor',
    '사이드&amp;커튼에어백' : 'side_airbag', 
    '운전석에어백' : 'driver_airbag', 
    '조수석에어백' : 'pass_airbag', 
    '후방카메라' : 'back_camera', 
    '어라운드뷰' : 'around_view', 
    '블랙박스' : 'blackbox', 
    'ABS' : 'ABS',
    'ECS' : 'ECS', 
    'TCS' : 'TCS', 
    '차체자세제어장치' : 'ESC', 
    '차선이탈경보장치' : 'LDWS', 
    '도난경보기' : 'bur_alarm', 
    '타이어 공기압감지' : 'TPMS', 
    '무릎에어백' : 'knee_airbag',
    '주차감지센서(전방)' : 'parking_sensor', 
    '전방카메라' : 'front_carmera', 
    '자동긴급제동(AEB)' : 'AEB', 
    '스마트키' : 'smartkey', 
    '파워윈도우' : 'power_window', 
    '자동도어잠금' : 'auto_doorlock',
    '풀오토에어컨' : 'fullauto_aircon', 
    '오토라이트' : 'auto_light', 
    '자동주차시스템' : 'auto_parking', 
    '전자식파킹브레이크' : 'elec_parking_break', 
    'HUD' : 'HUD', 
    'ECM룸밀러' : 'ECM', 
    '크루즈컨트롤' : 'cruise_control',
    '전동트렁크' : 'auto_trunk', 
    '핸들열선' : 'heated_steering', 
    '무선도어 잠금장치' : 'keyless_doorlock', 
    '레인센서와이퍼' : 'rain_sensor', 
    '스탑앤고' : 'stopNgo', 
    '무선충전' : 'wireless_charge', 
    '내비게이션' : 'navi',
    'CD플레이어' : 'CDrom',
    'CD체인저' : 'CD_changer', 
    'AV시스템' : 'av_sys',
    'MP3' : 'mp3', 
    'AUX' : 'AUX', 
    'USB' : 'USB', 
    '핸즈프리' : 'handsfree',
    'iPod 단자' : 'iPod',
    '뒷좌석모니터' : 'back_monitor', 
    '스마트폰미러링' : 'mirroring',
    '내비게이션 (비순정)' : 'non_genu_navi', 
    '블루투스' : 'bluetooth'
}

df = pd.read_csv('./car/tucson.csv')
df.rename(columns=mapper, inplace=True)
df.to_csv('./car/tucson_engColumns.csv', index=False)

for file in os.listdir('./car') :
#     print(file)
    df = pd.read_csv('./car/' + file)
    df.rename(columns=mapper, inplace=True)
    df.to_csv('./car/' + file[:-4] + '_engColumns.csv', index=False)

for file in os.listdir('./car') :
    if file.find('_') > 0 :
        df = pd.read_csv('./car/' + file)
        df.drop(columns='name', inplace=True)
        df.to_csv('./car/' + file[:-4] + '_noname.csv', index=False)
