==> 01. Preprocessing/03. Preprocessing_getDummies.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
import os

df_grandeur = pd.read_csv('./car/grandeur_firstName.csv')
df_k5 = pd.read_csv('./car/k5_firstName.csv')
df_morning = pd.read_csv('./car/morning_firstName.csv')
df_tucson = pd.read_csv('./car/tucson_firstName.csv')
print(df_grandeur.shape)
print(df_k5.shape)
print(df_morning.shape)
print(df_tucson.shape)

print(df_grandeur.isna().sum().sum())
print(df_k5.isna().sum().sum())
print(df_morning.isna().sum().sum())
print(df_tucson.isna().sum().sum())

df_grandeur

df_grandeur = pd.concat([pd.get_dummies(df_grandeur['first_name'], prefix='name'),
                         pd.get_dummies(df_grandeur['fuel'], prefix='fuel'),
                         pd.get_dummies(df_grandeur['color'], prefix='color'), 
                         df_grandeur], 
                        axis=1).drop(columns=['name', 'first_name', 'fuel', 'color'])

df_k5 = pd.concat([pd.get_dummies(df_k5['first_name'], prefix='name'),
                   pd.get_dummies(df_k5['fuel'], prefix='fuel'),
                   pd.get_dummies(df_k5['color'], prefix='color'), 
                   df_k5], 
                  axis=1).drop(columns=['name', 'first_name', 'fuel', 'color'])

df_morning = pd.concat([pd.get_dummies(df_morning['first_name'], prefix='name'),
                        pd.get_dummies(df_morning['fuel'], prefix='fuel'),
                        pd.get_dummies(df_morning['color'], prefix='color'), 
                        df_morning], 
                       axis=1).drop(columns=['name', 'first_name', 'fuel', 'color'])

df_tucson = pd.concat([pd.get_dummies(df_tucson['first_name'], prefix='name'),
                       pd.get_dummies(df_tucson['fuel'], prefix='fuel'),
                       pd.get_dummies(df_tucson['color'], prefix='color'), 
                       df_tucson], 
                      axis=1).drop(columns=['name', 'first_name', 'fuel', 'color'])

df_grandeur.columns

df_grandeur = df_grandeur[['year', 'distance', 'name_HG', 'name_IG', 'name_TG', 'name_normal', 'name_the_all_new',
       'name_the_luxury', 'name_the_new', 'fuel_LPG', 'fuel_diesel',
       'fuel_gasoline', 'fuel_hybrid', 'color_black', 'color_gray',
       'color_others', 'color_silver', 'color_white', 
       'sunroof', 'HID', 'Al_wheel', 'auto_mirror', 'DLR', 'sunroof_pano',
       'roof_rack', 'LED', 'HBA', 'ghost_door', 'sliding_door', 'side_step',
       'adap_h', 'driver_heated_seat', 'driver_ventilated_seat',
       'power_steering', 'steering_remote', 'leahter_seat', 'driver_elec_seat',
       'pass_elec_seat', 'back_elec_seat', 'driver_memory_seat',
       'massage_seat', 'supervision_cluster', 'paddle_shift', 'hi_pass',
       'pass_memory_seat', 'back_heated_seat', 'amb_light', 'walkin_seat',
       'sun_visor', 'pass_ventilated_seat', 'back_ventilated_seat',
       'back_sensor', 'side_airbag', 'driver_airbag', 'pass_airbag',
       'back_camera', 'around_view', 'blackbox', 'ABS', 'ECS', 'TCS', 'ESC',
       'LDWS', 'bur_alarm', 'TPMS', 'knee_airbag', 'parking_sensor',
       'front_carmera', 'AEB', 'smartkey', 'power_window', 'auto_doorlock',
       'fullauto_aircon', 'auto_light', 'auto_parking', 'elec_parking_break',
       'HUD', 'ECM', 'cruise_control', 'auto_trunk', 'heated_steering',
       'keyless_doorlock', 'rain_sensor', 'stopNgo', 'wireless_charge', 'navi',
       'CDrom', 'CD_changer', 'av_sys', 'mp3', 'AUX', 'USB', 'handsfree',
       'iPod', 'back_monitor', 'mirroring', 'non_genu_navi', 'bluetooth',
       'price']]

df_morning.columns

df_morning = df_morning[['year',
       'distance', 'name_all_new', 'name_all_new_van', 'name_new_L', 'name_new_LX',
       'name_new_SLX', 'name_normal', 'name_the_new', 'name_the_new_van',
       'fuel_LPG', 'fuel_gasoline', 'color_black', 'color_gray',
       'color_others', 'color_pearl', 'color_silver', 'color_white', 
       'sunroof', 'HID', 'Al_wheel', 'auto_mirror', 'DLR',
       'sunroof_pano', 'roof_rack', 'LED', 'HBA', 'ghost_door', 'sliding_door',
       'side_step', 'adap_h', 'driver_heated_seat', 'driver_ventilated_seat',
       'power_steering', 'steering_remote', 'leahter_seat', 'driver_elec_seat',
       'pass_elec_seat', 'back_elec_seat', 'driver_memory_seat',
       'massage_seat', 'supervision_cluster', 'paddle_shift', 'hi_pass',
       'pass_memory_seat', 'back_heated_seat', 'amb_light', 'walkin_seat',
       'sun_visor', 'pass_ventilated_seat', 'back_ventilated_seat',
       'back_sensor', 'side_airbag', 'driver_airbag', 'pass_airbag',
       'back_camera', 'around_view', 'blackbox', 'ABS', 'ECS', 'TCS', 'ESC',
       'LDWS', 'bur_alarm', 'TPMS', 'knee_airbag', 'parking_sensor',
       'front_carmera', 'AEB', 'smartkey', 'power_window', 'auto_doorlock',
       'fullauto_aircon', 'auto_light', 'auto_parking', 'elec_parking_break',
       'HUD', 'ECM', 'cruise_control', 'auto_trunk', 'heated_steering',
       'keyless_doorlock', 'rain_sensor', 'stopNgo', 'wireless_charge', 'navi',
       'CDrom', 'CD_changer', 'av_sys', 'mp3', 'AUX', 'USB', 'handsfree',
       'iPod', 'back_monitor', 'mirroring', 'non_genu_navi', 'bluetooth',
       'price']]

df_k5.columns

df_k5 = df_k5[['year',
       'distance', 'name_2nd', 'name_3rd', 'name_normal', 'name_the_new',
       'name_the_new_2nd', 'name_the_new_3rd', 'fuel_LPG', 'fuel_diesel',
       'fuel_gasoline', 'fuel_hybrid', 'color_black', 'color_gray',
       'color_others', 'color_pearl', 'color_silver', 'color_white', 'sunroof', 'HID', 'Al_wheel', 'auto_mirror', 'DLR',
       'sunroof_pano', 'roof_rack', 'LED', 'HBA', 'ghost_door', 'sliding_door',
       'side_step', 'adap_h', 'driver_heated_seat', 'driver_ventilated_seat',
       'power_steering', 'steering_remote', 'leahter_seat', 'driver_elec_seat',
       'pass_elec_seat', 'back_elec_seat', 'driver_memory_seat',
       'massage_seat', 'supervision_cluster', 'paddle_shift', 'hi_pass',
       'pass_memory_seat', 'back_heated_seat', 'amb_light', 'walkin_seat',
       'sun_visor', 'pass_ventilated_seat', 'back_ventilated_seat',
       'back_sensor', 'side_airbag', 'driver_airbag', 'pass_airbag',
       'back_camera', 'around_view', 'blackbox', 'ABS', 'ECS', 'TCS', 'ESC',
       'LDWS', 'bur_alarm', 'TPMS', 'knee_airbag', 'parking_sensor',
       'front_carmera', 'AEB', 'smartkey', 'power_window', 'auto_doorlock',
       'fullauto_aircon', 'auto_light', 'auto_parking', 'elec_parking_break',
       'HUD', 'ECM', 'cruise_control', 'auto_trunk', 'heated_steering',
       'keyless_doorlock', 'rain_sensor', 'stopNgo', 'wireless_charge', 'navi',
       'CDrom', 'CD_changer', 'av_sys', 'mp3', 'AUX', 'USB', 'handsfree',
       'iPod', 'back_monitor', 'mirroring', 'non_genu_navi', 'bluetooth',
       'price']]

df_tucson.columns

df_tucson = df_tucson[['year', 'distance', 'name_all_new', 'name_ix', 'name_new_ix', 'fuel_diesel',
       'fuel_gasoline', 'color_black', 'color_gray', 'color_others',
       'color_silver', 'color_white', 'sunroof', 'HID',
       'Al_wheel', 'auto_mirror', 'DLR', 'sunroof_pano', 'roof_rack', 'LED',
       'HBA', 'ghost_door', 'sliding_door', 'side_step', 'adap_h',
       'driver_heated_seat', 'driver_ventilated_seat', 'power_steering',
       'steering_remote', 'leahter_seat', 'driver_elec_seat', 'pass_elec_seat',
       'back_elec_seat', 'driver_memory_seat', 'massage_seat',
       'supervision_cluster', 'paddle_shift', 'hi_pass', 'pass_memory_seat',
       'back_heated_seat', 'amb_light', 'walkin_seat', 'sun_visor',
       'pass_ventilated_seat', 'back_ventilated_seat', 'back_sensor',
       'side_airbag', 'driver_airbag', 'pass_airbag', 'back_camera',
       'around_view', 'blackbox', 'ABS', 'ECS', 'TCS', 'ESC', 'LDWS',
       'bur_alarm', 'TPMS', 'knee_airbag', 'parking_sensor', 'front_carmera',
       'AEB', 'smartkey', 'power_window', 'auto_doorlock', 'fullauto_aircon',
       'auto_light', 'auto_parking', 'elec_parking_break', 'HUD', 'ECM',
       'cruise_control', 'auto_trunk', 'heated_steering', 'keyless_doorlock',
       'rain_sensor', 'stopNgo', 'wireless_charge', 'navi', 'CDrom',
       'CD_changer', 'av_sys', 'mp3', 'AUX', 'USB', 'handsfree', 'iPod',
       'back_monitor', 'mirroring', 'non_genu_navi', 'bluetooth', 'price']]

df_tucson

df_grandeur.to_csv('./car/grandeur_firstName_dummies.csv', index=False)
df_k5.to_csv('./car/k5_firstName_dummies.csv', index=False)
df_morning.to_csv('./car/morning_firstName_dummies.csv', index=False)
df_tucson.to_csv('./car/tucson_firstName_dummies.csv', index=False)
